src/magento_lib/inference/requirements.txt
PyYAML
gcsfs
numpy
pandas